renamed packages to get rid of warnings
