+++ routingengine/datasource.go
@@ -1,4 +1,4 @@
-package routing_engine
+package routingengine
 
 type RoutingEngineDatasource interface {
 	RouteEngineStats() (*RouteEngineStats, error)
